+++ packages/orq-rc/src/models/operations/postv2feedbackevaluation.ts
@@ -245,7 +245,7 @@ export const RequestBody4$outboundSchema: z.ZodType<
     PostV2FeedbackEvaluationRequestBodyFeedbackRequestSource$outboundSchema
       .default("orq"),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-09-03T10:05:19.694Z"))
+  reviewedAt: z.date().default(() => new Date("2026-09-03T11:12:45.130Z"))
     .transform(v => v.toISOString()),
   type: z.literal("string_array"),
   values: z.array(z.string()),
@@ -299,7 +299,7 @@ export const Three$outboundSchema: z.ZodType<
   source: PostV2FeedbackEvaluationRequestBodyFeedbackSource$outboundSchema
     .default("orq"),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-09-03T10:05:19.694Z"))
+  reviewedAt: z.date().default(() => new Date("2026-09-03T11:12:45.130Z"))
     .transform(v => v.toISOString()),
   type: z.literal("boolean"),
   value: z.boolean(),
@@ -353,7 +353,7 @@ export const RequestBody2$outboundSchema: z.ZodType<
     "orq",
   ),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-09-03T10:05:19.694Z"))
+  reviewedAt: z.date().default(() => new Date("2026-09-03T11:12:45.129Z"))
     .transform(v => v.toISOString()),
   type: z.literal("number"),
   value: z.number(),
@@ -405,7 +405,7 @@ export const RequestBody1$outboundSchema: z.ZodType<
   humanReviewId: z.string().optional(),
   source: RequestBodySource$outboundSchema.default("orq"),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-09-03T10:05:19.694Z"))
+  reviewedAt: z.date().default(() => new Date("2026-09-03T11:12:45.129Z"))
     .transform(v => v.toISOString()),
   type: z.literal("string"),
   value: z.string(),