+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3898,7 +3898,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M1KBNSA2EWCCHQYBN73S6NH9"),
+  id: z.string().default("01M1KFH7WD6QRZZDC0P3VB2D6M"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),