+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1764,7 +1764,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M1KBNSCC77NHPYFAZ7W2CVQB"),
+  _id: z.string().default("tool_01M1KFH7ZB1PGF4H7KFVTK9XNC"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -1959,7 +1959,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M1KBNSCBR82Z93DCAGPGSFTY"),
+  _id: z.string().default("tool_01M1KFH7ZAKDQB882YFRQ0SNTP"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2054,7 +2054,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M1KBNSCADJ4T3THA6HD8DQJS"),
+  _id: z.string().default("tool_01M1KFH7Z96FN3EF5R0W23B8RB"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2157,7 +2157,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M1KBNSCAX7MYNACX6JS524DV"),
+  _id: z.string().default("tool_01M1KFH7Z931WEF72BABTPHVF1"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),