+++ packages/orq-rc/RELEASES.md
@@ -12069,3 +12069,13 @@ Based on:
 - [typescript v4.15.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.17 - packages/orq-rc
+
+## 2026-09-03 11:15:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.3 (2.934.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.18 - packages/orq-rc